Namespace UiText::Entry use to avoid conflict with parent app
If the parent app has a model/class also named Entry, it can trigger a conflict where the UiText controller uses that class instead of the one from its namespace. This seems to happen mainly in development, where code is not eager loaded but have experienced a couple of instance where it fails in a running app.

This change ensures the entries controller uses the correct Entry class to avoid this conflict.